requirements.txt
paramiko==3.5.0
PyYAML==6.0
PyYAML==6.0.2
